% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_p_values.R
\name{calculate_p_values}
\alias{calculate_p_values}
\title{A function to calculate the p values of differences between sexes and attend/non-attend SAC.}
\usage{
calculate_p_values(drug_name, group, partition, data)
}
\arguments{
\item{drug_name}{The three letter code for the drug (e.g., pzq, alb, ...). In quotes.}

\item{group}{The code for the group of interest (e.g., "1_SAC", "2_Adults"). In quotes.}

\item{partition}{The variable to determine by what category the data will be split by. In quotes.}

\item{data}{The data from the survey. A dataframe, not in quotes.}
}
\value{
A data frame containing the p values for the different differences for the chosen drug and group.
}
\description{
A function to calculate the p values of differences between sexes and attend/non-attend SAC.
}
