% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_check.R
\name{checkingvariables}
\alias{checkingvariables}
\title{Cleaning checks}
\usage{
checkingvariables(data, variable, tosplitby, makefactor = NA)
}
\arguments{
\item{data}{The data to use. Not in quotes}

\item{variable}{The variable to use.formatted as data$variable - in quotes}

\item{tosplitby}{The variable to split the missings by - in quotes}

\item{makefactor}{if not NA, makes variable is turned to factor}
}
\value{
If the variable is a character or factor, the number of non-missing levels and
\itemize{
\item if there's less than 200 levels - the levels and how many rows in each
\item if there's more than 200 levels - the proportion of levels that are unique.
If the variables is numeric returns a summary
The number of missing variables broken down by 'tosplitby'
}
}
\description{
Function showing a summary of the variable and the number of missings for data cleaning
}
