% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dates.R
\name{date_unifier}
\alias{date_unifier}
\title{Extract dates from long date-time string}
\usage{
date_unifier(x)
}
\arguments{
\item{x}{Date and time field that is generated by SurveyCTO}
}
\value{
Date in the format YYYY-MM-DD
}
\description{
Legacy function for standardising dates autogenerated by SurveyCTO
}
\examples{
mutate(dataframe, mydate = date_unifier(SubmissionDate))
}
