% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drc_equity_tool_codes}
\alias{drc_equity_tool_codes}
\title{DRC Equity Tool Standard Questions.}
\format{
A data frame with 10 rows and 7 columns:
\describe{
\item{question}{Simplified code on what the question is about (\href{https://www.equitytool.org/democratic-republic-of-congo}{see} for more details)}
\item{op1}{Answer Option 1, text}
\item{op1_s}{Answer Option 1 score, numeric}
\item{op2}{Answer Option 2, text}
\item{op2_s}{Answer Option 2 score, numeric}
\item{op3}{Answer Option 3, text (not always available)}
\item{op3_s}{Answer Option 3 score, numeric (not always available)}
}
}
\source{
Simplified data frame based on the \href{https://www.equitytool.org/democratic-republic-of-congo}{Equity Tool Website's}
form for the Ivory Coast. Requires requesting Excel Workbook to be sent. Based on the content in the "Other software" sheet.
}
\usage{
drc_equity_tool_codes
}
\description{
A dataset containing the questions, answer codes and answer scores to the
\href{https://www.equitytool.org/development}{Equity Tool Website} set of questions for the Ivory Coast.
}
\keyword{datasets}
