% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equity_calc.R
\name{equity_calculation}
\alias{equity_calculation}
\title{Equity calcualtion}
\usage{
equity_calculation(data, vectorOfQuintiles, equity_file)
}
\arguments{
\item{data}{A data frame of HHs (more details)}

\item{vectorOfQuintiles}{A vector (more details)}

\item{equity_file}{A data frame (more details)}
}
\value{
A data frame containing the equity scores and the equity quintiles for each household of data (first parameter)
}
\description{
A function to calculate the equity values based on teh Equity Tool values.
}
