% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_method_cs_analysis.R
\name{estimate_cs_values}
\alias{estimate_cs_values}
\title{A function to obtain binary var (e.g., coverage: yes / no) estimates with CI from WHO method CS clean data.}
\usage{
estimate_cs_values(var, iu, design, level = 0.95, degf = NA)
}
\arguments{
\item{var}{A string denoting the name of the variable to analyse.}

\item{iu}{A string denoting the name of the variable to use for administrative disaggregation .}

\item{design}{A survey package design object.}

\item{level}{Confidence level for interval}

\item{degf}{Degrees of freedom. By default NA, which call uses the svyciprop default \code{defg(design)} evaluation (see svyciprop documentation). For user inputted degrees of freedom enter a positive integer.}
}
\value{
A data frame containing the survey estimates for the inputted variable over the inputted administrative disaggregation (commonly the implementation unit) for the survey groups (SAC and/or adults) over all individuals, disaggregated by sex and (for SAC) by school attendance.
}
\description{
A function to obtain binary var (e.g., coverage: yes / no) estimates with CI from WHO method CS clean data.
}
