% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_equity}
\alias{example_equity}
\title{Example Equity Tool Answers}
\format{
A data frame with 3282 rows and 10 columns:
\describe{
\item{electricity}{Does this household have electricity?}
\item{radio}{Does this household have a radio?}
\item{television}{Does this household have a television?}
\item{watch}{Does any member of this household own a watch?}
\item{mobile}{Does any member of this household own mobile phone?}
\item{drinking_water}{What is the main source of drinking water for members of your household?}
\item{ext_wall}{What is the primary material used in the construction of the exterior walls of your dwelling?}
\item{floor}{What is the primary material of the floor of your dwelling?}
\item{cooking_fuel}{What type of fuel does your household primarily use for cooking?}
\item{roof}{What is the primary material of the roof of your dwelling?}
}
}
\source{
Anonymised and amended based on survey data from the SCI Foundation.
}
\usage{
example_equity
}
\description{
A dataset containing the answers to the \href{https://www.equitytool.org/development}{Equity Tool}
set of questions for the DRC.
}
\keyword{datasets}
