% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_hh}
\alias{example_hh}
\title{Household anonymised data based on a survey of SAC from 3282 households in the DRC.}
\format{
A data frame with 3282 rows and 48 columns:
\describe{
\item{country_name}{t}
\item{country_code}{t}
\item{year_survey}{t}
\item{year_financial}{t}
\item{survey_code}{t}
\item{admin_1_name}{t}
\item{admin_1_label}{t}
\item{admin_1_code}{t}
\item{admin_2_name}{u}
\item{admin_2_label}{u}
\item{admin_2_code}{u}
\item{moh_1_name}{u}
\item{moh_1_label}{u}
\item{moh_1_code}{u}
\item{admin_iu}{u}
\item{segment_name}{u}
\item{segment_label}{u}
\item{segment_code}{u}
\item{segment_lat}{u}
\item{segment_lon}{u}
\item{segment_altitude}{u}
\item{segment_accuracy}{u}
\item{groups_surveyed}{u}
\item{pzq_survey_bin}{u}
\item{alb_survey_bin}{u}
\item{mbd_survey_bin}{u}
\item{ivm_survey_bin}{u}
\item{seg_segment_total}{u}
\item{seg_segment_surveyed}{u}
\item{hh_interviewer_init}{u}
\item{hh_house_number}{u}
\item{hh_house_code}{u}
\item{hh_lat}{u}
\item{hh_lon}{u}
\item{hh_altitude}{u}
\item{hh_accuracy}{u}
\item{hh_interviewed}{u}
\item{hh_reason_no_interview}{u}
\item{hh_consent_bin}{u}
\item{hh_hh_pop}{u}
\item{hh_adult_males}{u}
\item{hh_adult_females}{u}
\item{hh_sac_males}{u}
\item{hh_sac_females}{u}
\item{hh_eligible_pop_sac}{u}
\item{hh_eligible_pop_adult}{u}
\item{hh_national_equity_score}{u}
\item{hh_national_equity_quintile}{u}
}
}
\source{
Anonymised and amended based on survey data from the SCI Foundation.
}
\usage{
example_hh
}
\description{
A dataset containing household (hh) level information from a coverage survey in four health districts in Examplistan
following the \href{https://apps.who.int/iris/bitstream/handle/10665/329376/9789241516464-eng.pdf}{WHO CES Methodology}.
}
\keyword{datasets}
