% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_seg}
\alias{example_seg}
\title{Segment anonymised data based on a survey of SAC from 120 segments in the DRC.}
\format{
A data frame with 120 rows and 46 columns:
\describe{
\item{country_name}{t}
\item{country_code}{t}
\item{year_survey}{t}
\item{year_financial}{t}
\item{survey_code}{t}
\item{admin_1_name}{t}
\item{admin_1_label}{t}
\item{admin_1_code}{t}
\item{admin_2_name}{u}
\item{admin_2_label}{u}
\item{admin_2_code}{u}
\item{moh_1_name}{u}
\item{moh_1_label}{u}
\item{moh_1_code}{u}
\item{admin_iu}{u}
\item{segment_name}{u}
\item{segment_label}{u}
\item{segment_code}{u}
\item{segment_lat}{u}
\item{segment_lon}{u}
\item{segment_altitude}{u}
\item{segment_accuracy}{u}
\item{groups_surveyed}{u}
\item{pzq_survey_bin}{u}
\item{alb_survey_bin}{u}
\item{mbd_survey_bin}{u}
\item{ivm_survey_bin}{u}
\item{seg_interviewer_init}{u}
\item{seg_visit_date}{u}
\item{seg_int_position}{u}
\item{seg_population}{u}
\item{seg_no_hhs}{u}
\item{seg_pop_source}{u}
\item{seg_mda_type}{u}
\item{seg_date_mda}{u}
\item{seg_adults_treated_bin}{u}
\item{seg_community_how}{u}
\item{seg_how_door_to_door_bin}{u}
\item{seg_how_segment_head_bin}{u}
\item{seg_how_segment_centre_bin}{u}
\item{seg_how_segment_health_centre_bin}{u}
\item{seg_how_segment_school_bin}{u}
\item{seg_how_other_bin}{u}
\item{seg_how_no_knowledge_bin}{u}
\item{seg_segment_total}{u}
\item{seg_segment_surveyed}{u}
}
}
\source{
Anonymised and amended based on survey data from the SCI Foundation.
}
\usage{
example_seg
}
\description{
A dataset containing segment (seg) level information from a coverage survey in four health districts in Examplistan
following the \href{https://apps.who.int/iris/bitstream/handle/10665/329376/9789241516464-eng.pdf}{WHO CES Methodology}.
}
\keyword{datasets}
