% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_for_multiple_response_plots.R
\name{formultipleplot}
\alias{formultipleplot}
\title{Summarise a 'select multiple' variable for plotting}
\usage{
formultipleplot(data, ..., out)
}
\arguments{
\item{...}{Grouping variable(s)}

\item{out}{An outcome variable where more than one response can be selected}

\item{A}{dataframe containing grouping variables and outcome variable}
}
\value{
A summary tibble in long format for ggplot of n, total and percentage (n/total)
where n is frequency of each level within the outcome variable and total is the sum of n, grouped by grouping variable(s).
NAs are included because this function is often applied to follow-up questions to a binary question.
If NAs were excluded, reader won't know what the calculated percentage may be skewed to the readers as it becomes unclear
what the percentage is expressed on.
}
\description{
Function to summarise variable that allows multiple responses from CES and prepare for plotting
}
\examples{
When using function, need to specify out = otherwise function will assume it is a grouping variable
}
