% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_for_plots.R
\name{forplot}
\alias{forplot}
\title{Summarise a 'select one' variable for plotting}
\usage{
forplot(data, ..., out, dropna = c(FALSE, TRUE))
}
\arguments{
\item{data}{A dataframe containing grouping variables and outcome variable}

\item{...}{Grouping variable(s)}

\item{out}{An outcome variable}

\item{dropna}{Should NAs be dropped from dataframe?}
}
\value{
A summary tibble in long format for ggplot of n, total and percentage (n/total)
where n is frequency of each level within the outcome variable and total is the sum of n, grouped by grouping variable(s)
}
\description{
Function to summarise variables from CES and prepare for plotting
}
\examples{
When using function, need to specify out = otherwise function will assume it is a grouping variable
}
