% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelling_for_plots.R
\name{label_for_plot}
\alias{label_for_plot}
\title{Label variables for graphs}
\usage{
label_for_plot(dataset, ..., nchar = 20)
}
\arguments{
\item{dataset}{A dataframe containing the variables to be relabelled}

\item{...}{Variable(s) to be relabelled}

\item{nchar}{Maximum number of characters in each line (default is 20)}
}
\value{
An object of the same type as \code{dataset} with specified variable(s) as factor(s) with appropriate labelling.
}
\description{
Function for converting variable values to labels suitable for reporting and graphs.
It replaces all underscores with whitespaces and fixes whitespaces,
removes numbering options at the start of the string,
replaces the following abbreviations with the correct terms (CWT=community-wide, SBT=school-based, CDD=medicines distributors, Abd pain=abdominal),
It also converts variable into factor for plotting, and 'other' options are always ordered last.
}
\examples{
dataset \%>\% label_for_plot(vil_mda_type)
}
