% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_choice_dummies.R
\name{make_dummies}
\alias{make_dummies}
\title{Function that converts a multiple choice CES question into a dummy matrix of individual answers}
\usage{
make_dummies(data, variable, prefix)
}
\arguments{
\item{data}{The data to use. Not in quotes.}

\item{variable}{The variable with the multiple choice answers. In quotes.}

\item{prefix}{The prefix to use for the column names of the dummy matrix. In quotes.}
}
\value{
A data frame of the same length as the data frame passed as argument. Has as many columns as there
are unique answers in the variable to review.
}
\description{
Function that converts a multiple choice CES question into a dummy matrix of individual answers
}
