% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overall_df_plots.R
\name{make_overall_plot_df}
\alias{make_overall_plot_df}
\alias{make_overall_plot}
\alias{make_sci_theme}
\title{CES Overall plot data frame}
\usage{
make_overall_plot_df(survey_estimates, reported_df)

make_overall_plot(
  overall_plot_df,
  target = 0.75,
  target_text = "WHO\\nminimum target",
  theme_object = NA,
  y_axis_label_reach = "Programme Reach & 95\% CI (in \%)\\n",
  y_axis_label_coverage = "Survey Coverage & 95\% CI (in \%)\\n",
  x_axis_label = "\\nImplementation Unit",
  survey_point_shape = 16,
  survey_point_size = 4,
  reported_point_shape = 42,
  reported_point_size = 9,
  reported_text = "Reported coverage"
)

make_sci_theme()
}
\arguments{
\item{survey_estimates}{A data frame containing the output from the \code{estimate_cs_values} function.}

\item{reported_df}{A data frame on containing the reported coverage data. Needs to have the following four columns:
\itemize{
\item group - Denotes what group the reported data applies to (e.g., "SAC" or "Adult").
\item drug - Denotes the three letter abbreviation of the drug the reported data applies to (e.g., "PZQ").
\item partition - Denotes the partition the reported data applies to (e.g., "District X", "Commune Y").
\item estimate - Denotes the reported coverage data.
}}

\item{overall_plot_df}{Output from the \code{make_overall_plot_df}}

\item{target}{A numerice value between 0 and 1 to set the WHO target coverage. Default is 0.75.}

\item{target_text}{A string denoting the text that describes the target. Default is \verb{WHO\\nminimun target}.
Note the \verb{\\n} serves to make a new line in the legend text (see output).}

\item{theme_object}{A ggplot theme object (e.g. theme_classic()) for the plots. Defaults is NA.
In that case, the function calls \code{make_sci_theme()} to create the default theme. See that function for theme details.}

\item{y_axis_label_reach}{Y axis label for programme reach. Default is \verb{Programme Reach & 95\% CI (in \%)\\n}.
Note the \verb{\\n} is included to create a blank line underneath to make space. Same for other axis labels.}

\item{y_axis_label_coverage}{Y axis label for survey coverage. Default is \verb{Survey Coverage & 95\% CI (in \%)\\n}.}

\item{x_axis_label}{X axis label denoting the partitions. Default is \verb{Implementation Unit}}

\item{survey_point_shape}{Shape of the survey data point. Default is a filled dot (16).}

\item{survey_point_size}{Size of the survey data point. Default is 4.}

\item{reported_point_shape}{Shape of the reported coverage data point. Default is a 5 arms star (42).}

\item{reported_point_size}{Size of the reported coverage data point. Default is 9.}

\item{reported_text}{String describing the dots for reported coverage. Default is English: \verb{Reported coverage}.}
}
\value{
A data frame containing the information from the survey estimates and the reported data in a
format suitable to be passed to ggplot, in particular the \code{make_overall_plot} function.

A list of plots. Run names({make_overall_plot output}) to see them. Call them individually by using
\verb{\{make_overall_plot output\}[[\{name of the particular plot\}]]}.
}
\description{
A function to make a data frame to use for plotting for the overall cases (not disaggregated by sex or attend).

A function to make the overrall plot for dashboards with the \code{make_overall_plot_df} output.

Function to make SCI theme for plots if none passed to \code{make_overall_plot}.
}
