% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_response_plots.R
\name{multipleChoicePlot}
\alias{multipleChoicePlot}
\title{CES Multiple choice Plot
Function to summarise and plot graph of select_multiple/tick all that apply questions
Creates summary data and plots at two disaggregation levels
Also creates a simpler graph if necessary}
\usage{
multipleChoicePlot(
  data,
  var2work,
  firstLevelDisaggregation,
  secondLevelDisaggregation,
  thresholdInclude = 0,
  nFacetCols = 3,
  splittr = F,
  labels = vector(),
  simple = F,
  xLabel = "District",
  yLabel = "Percentage of children interviewed",
  rotate.x.labels = 0
)
}
\arguments{
\item{data}{A data frame of multiple choice responses}

\item{var2work}{A response variable to plot}

\item{firstLevelDisaggregation}{A primary grouping variable}

\item{secondLevelDisaggregation}{A secondary grouping variable}

\item{thresholdInclude}{A number of proportion 0-1}

\item{nFacetCols}{Number of facet columns}

\item{splittr}{If TRUE, splits facet_wrap variables. Defaults to FALSE.}

\item{simple}{If FALSE, adds geom_jitter. Defaults to FALSE}

\item{xLabel}{The x-axis label}

\item{yLabel}{The y-axis label}

\item{rotate.x.labels}{Rotates x-axis tick labels}
}
\value{
A list of plot and summary data
}
\description{
CES Multiple choice Plot
Function to summarise and plot graph of select_multiple/tick all that apply questions
Creates summary data and plots at two disaggregation levels
Also creates a simpler graph if necessary
}
