% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_response_plots.R
\name{proper}
\alias{proper}
\title{Converting a string to first letter capitalised, all other letters little.}
\usage{
proper(x, type = "cto")
}
\arguments{
\item{x}{A character vector}

\item{type}{String, either 'excel' or 'cto'}
}
\value{
The same character vector but every letter at the beginning of after a blank is capitalised.
If type = 'cto' also the two digits and all underscores are removed
}
\description{
This function does the same as the Excel function PROPER(string)
}
\examples{
proper(c("wOrds", "many words", "verymany many Words"))
[1] "Words"               "Many Words"          "Verymany Many Words"

proper(c( "01_farmer",  "02_merchant", "03_health_worker"))
[1] "Farmer"        "Merchant"      "Health Worker"

}
