% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting_tables.R
\name{rr_summary}
\alias{rr_summary}
\alias{rr_table}
\title{A function to calculate the p values of differences between sexes and attend/non-attend SAC.}
\usage{
rr_summary(partition, data)

rr_table(cs_values_output, p_values_output, rr_summary_output)
}
\arguments{
\item{partition}{The variable to determine by what category the data will be split by. In quotes.}

\item{data}{The data from the survey. A dataframe, not in quotes.}

\item{cs_values_output}{Data frame output of the \code{estimate_cs_values} function.}

\item{p_values_output}{Data frame output of the \code{calculate_p_values} function.}

\item{rr_summary_output}{Data frame output of the \code{rr_summary} function}
}
\value{
A data frame the number using the inputted data, broken up by the partition argument detailing:
\itemize{
\item the number of distinct segments visited,
\item the number of individuals interviewed,
\item the percentage ofdistinct segments, individuals interviewof containing the p values for the different differences for the chosen drug and group.
}

A data frame to b used in the SCIF RR for CES. Note that general info like \% of girls in the survey
is done once for every drug. Rarely is going to change, historically only one number is shown. Here kept
to show if something is off. When transcribing to Word or Excel it is easy to just drop that line.
}
\description{
A function to calculate the p values of differences between sexes and attend/non-attend SAC.

A function to create the RR table from the output of the other calculation functions.
}
