% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_legend.R
\name{shift_legend}
\alias{shift_legend}
\title{Function to shift legends into empty facets of a ggplot2 object
Ownership of this specific function belongs to Z.Lin
(Taken from https://stackoverflow.com/questions/54438495/shift-legend-into-empty-facets-of-a-faceted-plot-in-ggplot2)
As far as the authors of this package are aware, this specific function has not been published in a package.
Therefore, to aid data visualisation for works by SCIF, this function is compiled into sciCoverageR.}
\usage{
shift_legend(p)
}
\arguments{
\item{p}{A ggplot2 object or grob generated from ggplotGrob with empty facet}
}
\value{
A grob object to be plotted using grid.draw()
}
\description{
Function to shift legends into empty facets of a ggplot2 object
Ownership of this specific function belongs to Z.Lin
(Taken from https://stackoverflow.com/questions/54438495/shift-legend-into-empty-facets-of-a-faceted-plot-in-ggplot2)
As far as the authors of this package are aware, this specific function has not been published in a package.
Therefore, to aid data visualisation for works by SCIF, this function is compiled into sciCoverageR.
}
\examples{
library(ggplot2)
myplot <- ggplot(diamonds, aes(x = carat, fill = cut)) +
geom_density(position = "stack") +
facet_wrap(~ color)
library(grid)
grid.draw(shift_legend(myplot))
}
