% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation_to_french.R
\name{tofrench}
\alias{tofrench}
\title{Translate variables to french}
\usage{
tofrench(
  dataset,
  which = c("sens_how", "treatment_info", "resist", "benefits", "symptoms", "prevent",
    "recognise", "reason_no", "where_swallow", "decision", "how_taken", "vil_mda_type",
    "vil_community_how", "yesno", "gender", "school"),
  var
)
}
\arguments{
\item{dataset}{A dataframe containing variable to be translated}

\item{which}{Specifies which set of translations to match.
Individual questionnaire:
*\code{sens_how} for how they heard about MDA,
*\code{treatment_info} for what treatment info received,
*\code{resist},\code{benefits},\code{symptoms},\code{prevent} for what negative rumours/benefits/symptoms/prevention they were aware of
*\code{recognise} for recognition of drugs and terms,
*\code{reason_no} for reason not swallowed,
*\code{where_swallow} for where swallowed,
*\code{decision} for who decided,
*\code{how_taken} for how were drugs swallowed,
*\code{yesno} for binary yes/no questions,
*\code{gender} for male/female questions,
*\code{school} for education level and school type questions.
Village questionnaire:
*\code{vil_mda_type} for mda delivery platform,
*\code{vil_community_how} for community-wide delivery platform.}

\item{var}{The variable from dataframe to be translated}
}
\value{
An object of the same type as \code{dataset}, with additional column created named \code{var_fr} of the french translations.
}
\description{
Function for translating a fixed set of variables to french
}
\examples{
dataset \%>\% tofrench("yesno",ind_child_attendance)
will return `dataset` object with a new column named ind_child_attendance_fr, with values `Oui` and `Non`.
}
