% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlimithealth_colours.R
\docType{data}
\name{uh_colours}
\alias{uh_colours}
\title{list of Unlimit Health colours}
\format{
An object of class \code{character} of length 11.
}
\usage{
uh_colours
}
\description{
list of Unlimit Health colours
}
\keyword{datasets}
