% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlimithealth_colours.R
\name{uh_cols}
\alias{uh_cols}
\title{This returns colours in the colour list above as hex codes}
\usage{
uh_cols(...)
}
\arguments{
\item{...}{Names of uh_colours}
}
\description{
This returns colours in the colour list above as hex codes
}
\keyword{internal}
