% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlimithealth_colours.R
\name{uh_pal}
\alias{uh_pal}
\title{Function to interpolate palette}
\usage{
uh_pal(palette = "all", reverse = FALSE, skip, ...)
}
\arguments{
\item{palette}{Name of palette in uh_palettes}

\item{reverse}{Reserve the direction of the palette? Default is FALSE.}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Function to interpolate palette
}
\examples{
uh_pal(palette= "primary")(2)
}
\keyword{internal}
