% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlimithealth_colours.R
\docType{data}
\name{uh_palettes}
\alias{uh_palettes}
\title{Define palettes}
\format{
An object of class \code{list} of length 7.
}
\usage{
uh_palettes
}
\description{
Define palettes
}
\keyword{internal}
