% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlimithealth_colours.R
\docType{data}
\name{unlimithealth_colours}
\alias{unlimithealth_colours}
\alias{uh_cols}
\alias{uh_palettes}
\alias{uh_pal}
\alias{scale_color_uh}
\alias{scale_colour_uh}
\alias{scale_fill_uh}
\title{Custom Unlimit Health branded scale colour/fill function for ggplot2.}
\format{
An object of class \code{list} of length 7.
}
\usage{
uh_cols(...)

uh_palettes

uh_pal(palette = "all", reverse = FALSE, skip, ...)

scale_color_uh(palette = "all", discrete = TRUE, reverse = FALSE, skip, ...)

scale_colour_uh(palette = "all", discrete = TRUE, reverse = FALSE, skip, ...)

scale_fill_uh(palette = "all", discrete = TRUE, reverse = FALSE, skip, ...)
}
\arguments{
\item{...}{Other arguments passed on to ggplot2 scale function to control limits, breaks and labels etc.}

\item{palette}{A palette name from one of:
\itemize{
\item \code{all} - all Unlimit Health branded colours
\item \code{main} - the primary and secondary Unlimit Health colours (7)
\item \code{primary} - the 3 primary Unlimit Health colours
\item \code{secondary} - the 4 secondary Unlimit Health colours
\item \code{complementary} - the 4 highlight colours that are to complement & contrast the secondary palette, though similar in colours.
\item \code{yesno} - contrasting dark blue and neutral from the primary palette
\item \code{gender} - contrasting red earth and deep aqua from the secondary palette
}}

\item{reverse}{Reserve the direction of the palette? Default is FALSE.}

\item{skip}{A numeric vector. Index/Indices of colours in the palette to be skipped over.}

\item{discrete}{Default is TRUE. FALSE if a continuous colour gradient is required.}
}
\description{
Colours are assigned according to the order in the palette.
If variable has more levels than specified in palette, colourRamp to generate additional colours matching the palette.
The skip argument is optional but useful where some IUs are not treated
}
\examples{
uh_pal(palette= "primary")(2)
 scale_colour_uh(palette="main",skip = 3) skip is specified and the 3rd colour in the main palette will be skipped

}
\keyword{datasets}
