% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignette_hh}
\alias{vignette_hh}
\title{HH data from a survey of 2261 households in the Ivory Coast.}
\format{
A data frame with 2261 rows and 39 columns:
\describe{
\item{KEY_hh}{Individual HH KEY, identifies them uniquely}
\item{int_date}{Interview date, date, month, and year}
\item{segment}{Subunit segment surveyed}
\item{eligible}{u}
\item{sac_boys}{u}
\item{sac_girls}{u}
\item{hh_eligible}{u}
\item{no_interviews}{u}
\item{electricity}{u}
\item{television}{u}
\item{refrigerator}{u}
\item{canal}{u}
\item{cooker}{u}
\item{cd}{u}
\item{bank_account}{u}
\item{water}{u}
\item{toilet}{u}
\item{cooking_fuel}{u}
\item{floor}{u}
\item{hh_ind_visit_date}{u}
\item{hh_ind_submission}{u}
\item{hh_house_number}{u}
\item{hh_interviewed}{u}
\item{hh_reason_no_interview}{u}
\item{hh_consent}{u}
\item{hh_sac_male}{u}
\item{hh_sac_female}{u}
\item{hh_eligible_pop_child}{u}
\item{admin_2_name}{u}
\item{admin_2_label}{u}
\item{admin_2_code}{u}
\item{admin_1_name}{u}
\item{admin_1_label}{u}
\item{admin_1_code}{u}
\item{moh_1_name}{u}
\item{moh_1_label}{u}
\item{moh_1_code}{u}
\item{subunit_name}{u}
}
}
\source{
Anonymised and amended based on survey data from the SCI Foundation
}
\usage{
vignette_hh
}
\description{
A dataset containing Household (HH) level information from a coverage survey in three health districts in the Ivory Coast
following the \href{https://apps.who.int/iris/bitstream/handle/10665/329376/9789241516464-eng.pdf}{WHO CES Methodology}.
}
\keyword{datasets}
