% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vignette_vil}
\alias{vignette_vil}
\title{Village data from a survey of 90 subunits in 3 health districts in the Ivory Coast.}
\format{
A data frame with 90 rows and 32 columns:
\describe{
\item{int_position}{Text}
\item{vil_population}{Text}
\item{vil_no_hhs}{Text}
\item{vil_pop_source}{Text}
\item{know_date}{Text}
\item{community_how}{Text}
\item{vil_segment_total}{Text}
\item{KEY_vil}{Text}
\item{vil_visit_date}{Text}
\item{vil_submission}{Text}
\item{vil_mda_type}{Text}
\item{vil_date_mda}{Text}
\item{vil_adults_treated}{Text}
\item{vil_community_how}{Text}
\item{vil_how_door_to_door_bin}{Text}
\item{vil_how_village_head_bin}{Text}
\item{vil_how_village_centre_bin}{Text}
\item{vil_how_village_health_centre_bin}{Text}
\item{vil_how_village_school_bin}{Text}
\item{vil_how_other_bin}{Text}
\item{vil_how_no_knowledge_bin}{Text}
\item{vil_interviewer_init}{Text}
\item{admin_2_name}{Text}
\item{admin_1_name}{Text}
\item{moh_1_name}{Text}
\item{admin_2_code}{Text}
\item{admin_1_code}{Text}
\item{moh_1_code}{Text}
\item{admin_1_label}{Text}
\item{admin_2_label}{Text}
\item{moh_1_label}{Text}
\item{subunit_name}{Text}
}
}
\source{
Anonymised and amended based on survey data from the SCI Foundation.
}
\usage{
vignette_vil
}
\description{
A dataset containing village (vil) level information from a coverage survey in three health districts in the Ivory Coast
following the \href{https://apps.who.int/iris/bitstream/handle/10665/329376/9789241516464-eng.pdf}{WHO CES Methodology}.
}
\keyword{datasets}
