% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overall_df_plots.R
\name{make_overall_plot_df}
\alias{make_overall_plot_df}
\alias{make_sci_theme}
\title{CES Overall plot data frame}
\usage{
make_overall_plot_df(survey_estimates, reported_df)

make_sci_theme()
}
\arguments{
\item{survey_estimates}{A data frame containing the output from the \code{estimate_cs_values} function.}

\item{reported_df}{A data frame on containing the reported coverage data. Needs to have the following four columns:
\itemize{
\item group - Denotes what group the reported data applies to (e.g., "SAC" or "Adult").
\item drug - Denotes the three letter abbreviation of the drug the reported data applies to (e.g., "PZQ").
\item partition - Denotes the partition the reported data applies to (e.g., "District X", "Commune Y").
\item estimate - Denotes the reported coverage data.
}}
}
\value{
A data frame containing the information from the survey estimates and the reported data in a
format suitable to be passed to ggplot, in particular the \code{make_overall_plot} function.
}
\description{
A function to make a data frame to use for plotting for the overall cases (not disaggregated by sex or attend).

Function to make SCI theme for plots if none passed to \code{make_overall_plot}.
}
